requirements.txt
# === Wheel ===
wheel>=0.42.0

# === Core Web App ===
flask==2.3.3
werkzeug==2.3.7
markdown==3.4.4

# === AI & NLP ===

transformers>=4.39.0
sentence-transformers==2.7.0
huggingface_hub[hf_xet]
huggingface_hub>=0.17.0
torch>=2.3.0

# === LangChain (Compatible Set) ===
langchain>=0.1.30
langchain-core>=0.2.43
langchain-community>=0.0.40
langchain-text-splitters>=0.0.2
langchain-huggingface==0.0.3


# === Vector DB ===
chromadb==0.4.24

# === LoRA / LLM Compile ===
llama-cpp-python>=0.2.32,<0.3.0
cmake>=3.25.0

# === Utility ===
striprtf>=0.0.26,<0.0.27
python-docx==0.8.11
typing-extensions>=4.8.0
numpy>=1.26.0,<2.0
scikit-learn==1.3.0
aiofiles==23.2.1
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.29.0
psutil>=5.9.8
tiktoken>=0.5.1,<1.0.0
packaging<25,>=23.2
faiss-cpu>=1.11.0
jsonschema>=4.19.0,<5.0.0
genson>=1.2.2

# === Llama Index ===
llama-index>=0.11.22
llama-index-core>=0.11.22
llama-index-readers-file>=0.2.1
llama-index-embeddings-huggingface>=0.1.5
llama-index-llms-llama-cpp>=0.2.2


# === TTS + Audio ===
pyttsx3==2.90
sounddevice==0.4.6
pydub==0.25.1
